+++ composeApp/src/iosMain/kotlin/com/nuvio/app/features/player/PlayerSettingsStorage.ios.kt
@@ -486,6 +486,7 @@ actual object PlayerSettingsStorage {
     actual fun saveHwAccelEnabled(enabled: Boolean) {}
     actual fun loadDiscordRichPresenceEnabled(): Boolean? = null
     actual fun saveDiscordRichPresenceEnabled(enabled: Boolean) {}
+    actual fun discordRichPresenceUnavailableReason(): String? = null
     actual fun loadTrayIconEnabled(): Boolean? = null
     actual fun saveTrayIconEnabled(enabled: Boolean) {}
     actual fun loadAudioOutput(): String? = null